refactor: organize tests imports
