+++ tests/domain/mocks/mock-account.ts
@@ -1,6 +1,5 @@
 import { faker } from '@faker-js/faker'
-import { type AddAccount } from '@/domain/usecases/authentication/add-account'
-import { type Authentication } from '@/domain/usecases/authentication/authentication'
+import { type AddAccount, type Authentication } from '@/domain/usecases/authentication'
 
 export const mockAddAccountInput = (): AddAccount.Input => ({
   username: faker.person.firstName(),
