+++ tests/infra/cryptography/bcrypt-adapter.spec.ts
@@ -1,6 +1,6 @@
 import bcrypt from 'bcrypt'
-import { BcryptAdapter } from '@/infra/cryptography'
 import { throwError } from '@/tests/domain/mocks'
+import { BcryptAdapter } from '@/infra/cryptography'
 
 jest.mock('bcrypt', () => ({
   async hash(): Promise<string> {
