+++ tests/infra/db/mongodb/static/authentication/static-account-mongo-repository.spec.ts
@@ -1,5 +1,5 @@
-import { faker } from '@faker-js/faker'
 import { Collection } from 'mongodb'
+import { faker } from '@faker-js/faker'
 import { mockAddAccountInput, throwError } from '@/tests/domain/mocks'
 import { StaticAccountMongoRepository } from '@/infra/db/mongodb/static/authentication'
 import { MongoHelper } from '@/infra/db/mongodb'
