+++ tests/main/factories/validations/authentication/sign-up-validation-factory.spec.ts
@@ -1,7 +1,7 @@
+import { makeSignUpValidation } from '@/main/factories/validations/authentication'
 import { ValidationComposite, EmailValidation, CompareFieldsValidation, RequiredFieldValidation } from '@/validation/validators'
 import { type Validation } from '@/presentation/protocols'
 import { EmailValidatorAdapter } from '@/infra/validators'
-import { makeSignUpValidation } from '@/main/factories/validations/authentication'
 
 jest.mock('@/validation/validators/validation-composite')
