+++ tests/main/middlewares/content-type.test.ts
@@ -1,6 +1,6 @@
 import { type Express } from 'express'
 import request from 'supertest'
-import { setupApp } from '@/main/config/app'
+import { setupApp } from '@/main/config'
 
 let app: Express
