+++ tests/main/routes/authentication-route.test.ts
@@ -1,9 +1,9 @@
 import { type Express } from 'express'
 import { type Collection } from 'mongodb'
-import request from 'supertest'
-import { MongoHelper } from '@/infra/db/mongodb'
-import { setupApp } from '@/main/config'
 import { hash } from 'bcrypt'
+import request from 'supertest'
+import { setupApp } from '@/main/config'
+import { MongoHelper } from '@/infra/db/mongodb'
 
 let accountCollection: Collection
 let app: Express
