+++ tests/presentation/controllers/authentication/sign-up-controller.spec.ts
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker'
 import { ValidationSpy, AddAccountSpy, AuthenticationSpy } from '@/tests/presentation/mocks'
-import { throwError } from '@/tests/domain/mocks/test-helper'
+import { throwError } from '@/tests/domain/mocks'
 import { SignUpController } from '@/presentation/controllers/authentication'
 import { HttpHelper } from '@/presentation/helpers'
 import { EmailInUseError, ServerError } from '@/presentation/errors'
